fix: 优化插件卸载时对 Worker 的处理 | Optimize the processing of Worker when the plugin is unloading.

+++ src/index.ts
@@ -129,12 +129,17 @@ export default class WakaTimePlugin extends siyuan.Plugin {
         this.eventBus.off("loaded-protyle", this.loadedProtyleEventListener);
         this.eventBus.off("click-editorcontent", this.clickEditorContentEventListener);
 
-        this.bridge
-            ?.call<THandlers["unload"]>("unload")
-            .then(() => {
-                this.bridge?.terminate();
-                this.worker?.terminate();
-            });
+        if (this.worker) {
+            this.bridge
+                ?.call<THandlers["unload"]>("unload")
+                .then(() => {
+                    this.bridge?.terminate();
+                    this.worker?.terminate();
+                });
+        }
+        else {
+            this.bridge?.terminate();
+        }
     }
 
     openSetting(): void {
